Make the azure pipelines really test new contaiers.
Currently, in the case of a failure to build an image,the "test" to check buildconf just causes the pipeline to download the latest successful version from dockerhub

If we create a dummy named container just for the purposes of running that sanity check then there won't be a container on dockerhub which might not wor

I expect this PR will expose quite a few (likely ffmpeg v3.2) containers where the build has been broken for a while now but the pipeline _appears_ to succeed.

Let's just preempt some of the 3.2 variants that are definitely going to fail as a result of recent ubuntu 16.04 -> 18.04 and alpine 3.8 -> 3.11 version bumps.